Correção de problema ao mostrar cliente na locacao

+++ crm/controller.py
@@ -264,39 +264,40 @@ def set_pay_order(work):
 
     month_quantity = diff_month(work.end_date, start_date)
     month_pay = list(rrule(freq=MONTHLY, count=month_quantity, dtstart=start_date))
+    if len(month_pay) > 0:
 
-    #Se o primeiro vencimento for no passado ou hoje, altera para o proximo mes
-    if datetime.now().day >= int(work.pay_date) or start_date.day >= int(work.pay_date):
-        start_date = month_pay[0].replace(day=1).replace(month=month_pay[0].month + 1)
-        month_pay = list(rrule(freq=MONTHLY, count=month_quantity, dtstart=start_date))
-    date_list = []
+        #Se o primeiro vencimento for no passado ou hoje, altera para o proximo mes
+        if datetime.now().day >= int(work.pay_date) or start_date.day >= int(work.pay_date):
+            start_date = month_pay[0].replace(day=1).replace(month=month_pay[0].month + 1)
+            month_pay = list(rrule(freq=MONTHLY, count=month_quantity, dtstart=start_date))
+        date_list = []
 
-    for i in month_pay:
-        max_day = calendar.monthrange(i.year,i.month)[1]
-        value = work.month_value
-        if int(max_day) < int(work.pay_date):
-            i = i.replace(day=max_day)
-        else:
-            i = i.replace(day=int(work.pay_date))
-        if i.date() <= work.end_date:
-            if PayOrder.objects.filter(work=work, customer=work.customer,
-                                       due_date__year=i.year, due_date__month=i.month).exists():
-                pay_order = PayOrder.objects.get(work=work, customer=work.customer, due_date__year=i.year,
-                                                     due_date__month=i.month)
-                if pay_order.status == PAY_ORDER_STATUS_PENDING:
-                    if pay_order.due_date >= datetime.now().date():
-                        pay_order.value = value
-                        pay_order.due_date = i
-                        pay_order.save()
+        for i in month_pay:
+            max_day = calendar.monthrange(i.year,i.month)[1]
+            value = work.month_value
+            if int(max_day) < int(work.pay_date):
+                i = i.replace(day=max_day)
             else:
-                PayOrder.objects.update_or_create(work=work, customer=work.customer, due_date=i,
-                                                  defaults={'value': value})
+                i = i.replace(day=int(work.pay_date))
+            if i.date() <= work.end_date:
+                if PayOrder.objects.filter(work=work, due_date__year=i.year, due_date__month=i.month).exists():
+                    pay_order = PayOrder.objects.get(work=work, due_date__year=i.year,
+                                                         due_date__month=i.month)
+                    if pay_order.status == PAY_ORDER_STATUS_PENDING:
+                        if pay_order.due_date >= datetime.now().date():
+                            pay_order.value = value
+                            pay_order.due_date = i
+                            pay_order.customer = work.customer
+                            pay_order.save()
+                else:
+                    PayOrder.objects.update_or_create(work=work, due_date=i,
+                                                      defaults={'value': value, 'customer': work.customer})
 
-            date_list.append(i)
+                date_list.append(i)
 
-    #exclui pagamentos antigos que não serão efetivados
-    PayOrder.objects.filter(work=work, status=PAY_ORDER_STATUS_PENDING,
-                            due_date__gt=datetime.now()).exclude(due_date__in=date_list).delete()
+        #exclui pagamentos antigos que não serão efetivados
+        PayOrder.objects.filter(work=work, status=PAY_ORDER_STATUS_PENDING,
+                                due_date__gt=datetime.now()).exclude(due_date__in=date_list).delete()
 
     return work
